refactor: pins 에 generic 적용
