+++ src/main/java/bowling/domain/frame/FinalFrameScore.java
@@ -3,20 +3,26 @@ package bowling.domain.frame;
 import bowling.domain.pin.FinalPins;
 import bowling.domain.pin.Pin;
 
+import java.util.Objects;
+
 public final class FinalFrameScore extends FrameScore {
 
     public static final int MAX_PIN_COUNT = 10;
+
+    private final FinalPins pins;
 
     public FinalFrameScore() {
         this(new FinalPins());
     }
 
     public FinalFrameScore(FinalPins pins) {
+
         this(pins, pins.frameStatus());
     }
 
     private FinalFrameScore(FinalPins pins, FrameStatus frameStatus) {
-        super(pins, frameStatus);
+        super(frameStatus);
+        this.pins = pins;
     }
 
     @Override
@@ -24,7 +30,7 @@ public final class FinalFrameScore extends FrameScore {
         if (frameStatus == FrameStatus.NOT_ENDED) {
             return EMPTY_STRING;
         }
-        final FinalPins finalPins = (FinalPins) pins;
+        final FinalPins finalPins = pins;
         final Pin firstPin = finalPins.firstPin();
         final Pin secondPin = finalPins.secondPin();
         final Pin thirdPin = finalPins.thirdPin();
@@ -41,6 +47,20 @@ public final class FinalFrameScore extends FrameScore {
 
     @Override
     public FrameScore knockDownPin(Pin pin) {
-        return new FinalFrameScore((FinalPins) pins.knockDownPin(pin));
+        return new FinalFrameScore(pins.knockDownPin(pin));
+    }
+
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        FinalFrameScore that = (FinalFrameScore) o;
+        return Objects.equals(pins, that.pins) && frameStatus == that.frameStatus;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(pins, frameStatus);
     }
 }
