+++ src/main/java/bowling/domain/frame/FrameScore.java
@@ -1,9 +1,6 @@
 package bowling.domain.frame;
 
 import bowling.domain.pin.Pin;
-import bowling.domain.pin.Pins;
-
-import java.util.Objects;
 
 public abstract class FrameScore {
 
@@ -12,11 +9,10 @@ public abstract class FrameScore {
     public static final String STRIKE_PIN_STATUS = "X";
     public static final String EMPTY_STRING = "";
 
-    protected final Pins pins;
+
     protected final FrameStatus frameStatus;
 
-    protected FrameScore(Pins pins, FrameStatus frameStatus) {
-        this.pins = pins;
+    protected FrameScore(FrameStatus frameStatus) {
         this.frameStatus = frameStatus;
     }
 
@@ -28,16 +24,4 @@ public abstract class FrameScore {
         return FrameStatus.NOT_ENDED != frameStatus;
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        FrameScore that = (FrameScore) o;
-        return Objects.equals(pins, that.pins) && frameStatus == that.frameStatus;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(pins, frameStatus);
-    }
 }
