+++ src/main/java/bowling/domain/frame/NormalFrameScore.java
@@ -2,20 +2,24 @@ package bowling.domain.frame;
 
 import bowling.domain.pin.NormalPins;
 import bowling.domain.pin.Pin;
-import bowling.domain.pin.Pins;
+
+import java.util.Objects;
 
 public final class NormalFrameScore extends FrameScore {
+
+    private final NormalPins pins;
 
     public NormalFrameScore() {
         this(new NormalPins());
     }
 
-    public NormalFrameScore(Pins pins) {
+    public NormalFrameScore(NormalPins pins) {
         this(pins, pins.frameStatus());
     }
 
-    public NormalFrameScore(Pins pins, FrameStatus frameStatus) {
-        super(pins, frameStatus);
+    public NormalFrameScore(NormalPins pins, FrameStatus frameStatus) {
+        super(frameStatus);
+        this.pins = pins;
     }
 
     @Override
@@ -34,6 +38,19 @@ public final class NormalFrameScore extends FrameScore {
 
     @Override
     public FrameScore knockDownPin(Pin pin) {
-        return new NormalFrameScore(pins.knockDownPin(pin));
+        return new NormalFrameScore(this.pins.knockDownPin(pin));
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        NormalFrameScore that = (NormalFrameScore) o;
+        return Objects.equals(pins, that.pins) && frameStatus == that.frameStatus;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(pins, frameStatus);
     }
 }
