+++ src/test/java/bowling/domain/frame/NormalFrameScoreTest.java
@@ -2,7 +2,6 @@ package bowling.domain.frame;
 
 import bowling.domain.pin.NormalPins;
 import bowling.domain.pin.Pin;
-import bowling.domain.pin.Pins;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -15,7 +14,7 @@ class NormalFrameScoreTest {
     @DisplayName("Pins를 입력받아 FrameScore를 생성한다.")
     void create() {
         // given
-        final Pins pins = new NormalPins(new Pin(), new Pin(0));
+        final NormalPins pins = new NormalPins(new Pin(), new Pin(0));
 
         // when
         final FrameScore frameScore = new NormalFrameScore(pins);
